requirements.txt
python-telegram-bot
python-dotenv
requests
geopy
nest_asyncio